Fixes #30: Vastly improved escaping implementation.
After some discussion it was determined that we should just be escaping the message body when it is being sent to Slack itself, not before it is saved into the database.  This change eliminates all of those steps and adds a public method to the MessageSender class that takes care of this.