Connects #76:  Initial commit for time zone and time selection support within the Dolores Landingham bot.
This changeset is the beginning of support for being able to set the time zone for an employee and select a time at which to send a scheduled message out in addition to the day after a given start date.  The intention is that the scheduled message will respect the time zone of an employee and be sent out according to their local time.

Example:

A scheduled message is set to be sent the 1st day after an employee starts at 10:00 AM.  Two new hires join, one in the EST time zone and one in the PST time zone.  The message should be sent at 10:00 AM to both employees at their local time, not at 10:00 AM EST and 7:00 AM PST.